feature flag audit logs backend changes

+++ backend/packages/Upgrade/src/api/controllers/FeatureFlagController.ts
@@ -1,4 +1,4 @@
-import { JsonController, Authorized, Post, Body, Delete, Put, Req, Get, Params, Patch, Res } from 'routing-controllers';
+import { JsonController, Authorized, Post, Body, Delete, Put, Req, Get, Params, Patch, Res, CurrentUser } from 'routing-controllers';
 import { FeatureFlagService } from '../services/FeatureFlagService';
 import { FeatureFlag } from '../models/FeatureFlag';
 import { FeatureFlagSegmentExclusion } from '../models/FeatureFlagSegmentExclusion';
@@ -7,12 +7,15 @@ import { FeatureFlagStatusUpdateValidator } from './validators/FeatureFlagStatus
 import { FeatureFlagFilterModeUpdateValidator } from './validators/FeatureFlagFilterModeUpdateValidator';
 import { FeatureFlagPaginatedParamsValidator } from './validators/FeatureFlagsPaginatedParamsValidator';
 import { AppRequest, PaginationResponse } from '../../types';
-import { SERVER_ERROR } from 'upgrade_types';
+import { EXPERIMENT_LOG_TYPE, FEATURE_FLAG_LIST_FILTER_MODE, FeatureFlagDeletedData, SERVER_ERROR } from 'upgrade_types';
 import { FeatureFlagValidation, IdValidator, UserParamsValidator } from './validators/FeatureFlagValidator';
 import { ExperimentUserService } from '../services/ExperimentUserService';
 import { FeatureFlagListValidator } from '../controllers/validators/FeatureFlagListValidator';
 import { Segment } from 'src/api/models/Segment';
 import { Response } from 'express';
+import { User } from '../models/User';
+import { ExperimentAuditLogRepository } from '../repositories/ExperimentAuditLogRepository';
+import { InjectRepository } from 'typeorm-typedi-extensions';
 
 interface FeatureFlagsPaginationInfo extends PaginationResponse {
   nodes: FeatureFlag[];
@@ -113,7 +116,7 @@ interface FeatureFlagsPaginationInfo extends PaginationResponse {
 @Authorized()
 @JsonController('/flags')
 export class FeatureFlagsController {
-  constructor(public featureFlagService: FeatureFlagService, public experimentUserService: ExperimentUserService) {}
+  constructor(@InjectRepository() private experimentAuditLogRepository: ExperimentAuditLogRepository, public featureFlagService: FeatureFlagService, public experimentUserService: ExperimentUserService) {}
 
   /**
    * @swagger
@@ -297,9 +300,10 @@ export class FeatureFlagsController {
   @Post()
   public create(
     @Body({ validate: true }) flag: FeatureFlagValidation,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<FeatureFlag> {
-    return this.featureFlagService.create(flag, request.logger);
+    return this.featureFlagService.create(flag, currentUser, request.logger);
   }
 
   /**
@@ -312,7 +316,7 @@ export class FeatureFlagsController {
    *       parameters:
    *         - in: body
    *           name: statusUpdate
-   *           description: Updating the featur flag's status
+   *           description: Updating the feature flag's status
    *           schema:
    *             type: object
    *             required:
@@ -335,9 +339,10 @@ export class FeatureFlagsController {
   @Patch('/status')
   public async updateState(
     @Body({ validate: true })
-    flag: FeatureFlagStatusUpdateValidator
+    flag: FeatureFlagStatusUpdateValidator,
+    @CurrentUser() currentUser: User,
   ): Promise<FeatureFlag> {
-    return this.featureFlagService.updateState(flag.flagId, flag.status);
+    return this.featureFlagService.updateState(flag.flagId, flag.status, currentUser);
   }
 
   /**
@@ -350,7 +355,7 @@ export class FeatureFlagsController {
    *       parameters:
    *         - in: body
    *           name: updateFilterMode
-   *           description: Updating the featur flag's filter mode
+   *           description: Updating the feature flag's filter mode
    *           schema:
    *             type: object
    *             required:
@@ -373,9 +378,10 @@ export class FeatureFlagsController {
   @Patch('/filterMode')
   public async updateFilterMode(
     @Body({ validate: true })
-    flag: FeatureFlagFilterModeUpdateValidator
+    flag: FeatureFlagFilterModeUpdateValidator,
+    @CurrentUser() currentUser: User,
   ): Promise<FeatureFlag> {
-    return this.featureFlagService.updateFilterMode(flag.flagId, flag.filterMode);
+    return this.featureFlagService.updateFilterMode(flag.flagId, flag.filterMode, currentUser);
   }
 
   /**
@@ -402,9 +408,10 @@ export class FeatureFlagsController {
   @Delete('/:id')
   public delete(
     @Params({ validate: true }) { id }: IdValidator,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<FeatureFlag | undefined> {
-    return this.featureFlagService.delete(id, request.logger);
+    return this.featureFlagService.delete(id, currentUser, request.logger);
   }
 
   /**
@@ -441,9 +448,10 @@ export class FeatureFlagsController {
     @Params({ validate: true }) { id }: IdValidator,
     @Body({ validate: true })
     flag: FeatureFlagValidation,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<FeatureFlag> {
-    return this.featureFlagService.update(flag, request.logger);
+    return this.featureFlagService.update(flag, currentUser, request.logger);
   }
 
   /**
@@ -455,7 +463,7 @@ export class FeatureFlagsController {
    *         - application/json
    *       parameters:
    *         - in: body
-   *           name: addinclusionList
+   *           name: addInclusionList
    *           description: Adding an inclusion list to the feature flag
    *           schema:
    *             type: object
@@ -471,9 +479,10 @@ export class FeatureFlagsController {
   @Post('/inclusionList')
   public async addInclusionList(
     @Body({ validate: true }) inclusionList: FeatureFlagListValidator,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<FeatureFlagSegmentInclusion> {
-    return this.featureFlagService.addList(inclusionList, 'inclusion', request.logger);
+    return this.featureFlagService.addList(inclusionList, FEATURE_FLAG_LIST_FILTER_MODE.INCLUSION, currentUser, request.logger);
   }
 
   /**
@@ -501,9 +510,10 @@ export class FeatureFlagsController {
   @Post('/exclusionList')
   public async addExclusionList(
     @Body({ validate: true }) exclusionList: FeatureFlagListValidator,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<FeatureFlagSegmentExclusion> {
-    return this.featureFlagService.addList(exclusionList, 'exclusion', request.logger);
+    return this.featureFlagService.addList(exclusionList, FEATURE_FLAG_LIST_FILTER_MODE.EXCLUSION, currentUser, request.logger);
   }
 
   /**
@@ -538,6 +548,7 @@ export class FeatureFlagsController {
   public async updateExclusionList(
     @Params({ validate: true }) { id }: IdValidator,
     @Body({ validate: true }) exclusionList: FeatureFlagListValidator,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<FeatureFlagSegmentExclusion> {
     if (id !== exclusionList.list.id) {
@@ -547,7 +558,7 @@ export class FeatureFlagsController {
         )
       );
     }
-    return this.featureFlagService.updateList(exclusionList, 'exclusion', request.logger);
+    return this.featureFlagService.updateList(exclusionList, FEATURE_FLAG_LIST_FILTER_MODE.EXCLUSION, currentUser, request.logger);
   }
 
   /**
@@ -582,6 +593,7 @@ export class FeatureFlagsController {
   public async updateInclusionList(
     @Params({ validate: true }) { id }: IdValidator,
     @Body({ validate: true }) inclusionList: FeatureFlagListValidator,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<FeatureFlagSegmentInclusion> {
     if (id !== inclusionList.list.id) {
@@ -591,7 +603,7 @@ export class FeatureFlagsController {
         )
       );
     }
-    return this.featureFlagService.updateList(inclusionList, 'inclusion', request.logger);
+    return this.featureFlagService.updateList(inclusionList, FEATURE_FLAG_LIST_FILTER_MODE.INCLUSION, currentUser, request.logger);
   }
 
   /**
@@ -619,9 +631,10 @@ export class FeatureFlagsController {
   @Delete('/inclusionList/:id')
   public async deleteInclusionList(
     @Params({ validate: true }) { id }: IdValidator,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<Segment> {
-    return this.featureFlagService.deleteList(id, request.logger);
+    return this.featureFlagService.deleteList(id, FEATURE_FLAG_LIST_FILTER_MODE.INCLUSION, currentUser, request.logger);
   }
 
   /**
@@ -649,9 +662,10 @@ export class FeatureFlagsController {
   @Delete('/exclusionList/:id')
   public async deleteExclusionList(
     @Params({ validate: true }) { id }: IdValidator,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest
   ): Promise<Segment> {
-    return this.featureFlagService.deleteList(id, request.logger);
+    return this.featureFlagService.deleteList(id, FEATURE_FLAG_LIST_FILTER_MODE.INCLUSION, currentUser, request.logger);
   }
 
   /**
@@ -683,6 +697,7 @@ export class FeatureFlagsController {
   @Get('/export/:id')
   public async exportFeatureFlag(
     @Params({ validate: true }) { id }: IdValidator,
+    @CurrentUser() currentUser: User,
     @Req() request: AppRequest,
     @Res() response: Response
   ): Promise<Response> {
@@ -691,6 +706,15 @@ export class FeatureFlagsController {
     response.setHeader('Content-Disposition', `attachment; filename="${featureFlag.name}.json"`);
     response.setHeader('Content-Type', 'application/json');
     const plainFeatureFlag = JSON.stringify(featureFlag, null, 2); // Convert to JSON string
+    const exportAuditLog: FeatureFlagDeletedData = {
+      flagName: featureFlag.name,
+    };
+
+    await this.experimentAuditLogRepository.saveRawJson(
+      EXPERIMENT_LOG_TYPE.FEATURE_FLAG_DESIGN_EXPORTED,
+      exportAuditLog,
+      currentUser,
+    );
     return response.send(plainFeatureFlag);
   }
 }
